requirements.txt
numpy~=2.2.6
joblib~=1.5.2
tensorflow~=2.20.0
opencv-python~=4.12.0.88
kaggle~=1.7.4.5
scikit-image~=0.25.2
python-dotenv~=1.0.0